Revert "[omaha-client] Use config capabilities"
This reverts commit b096f9762040bc1df2af7398cd4e8ae82c99dcd9.

Reason for revert: Broke omaha client e2e tests
Original change's description:
> [omaha-client] Use config capabilities
>
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1434515
> Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
> Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1458048
Exempt: RubberStamper 🤖 <android-build-ayeaye@system.gserviceaccount.com>
Commit-Queue: John Wittrock <wittrock@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>

+++ src/lib/assembly/config_schema/src/platform_settings/swd_config.rs
@@ -103,10 +103,10 @@ type ServerUrl = String;
 #[serde(default, deny_unknown_fields)]
 pub struct PolicyConfig {
     pub allow_reboot_when_idle: bool,
-    pub startup_delay_seconds: u16,
-    pub periodic_interval_minutes: u16,
-    pub fuzz_percentage_range: u8,
-    pub retry_delay_seconds: u16,
+    pub startup_delay_seconds: u32,
+    pub periodic_interval_minutes: u32,
+    pub fuzz_percentage_range: u32,
+    pub retry_delay_seconds: u32,
 }
 
 impl Default for PolicyConfig {
